backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
firebase-admin==6.5.0
google-cloud-aiplatform==1.70.0
google-cloud-firestore==2.18.0
google-adk>=0.1.0
httpx==0.27.0
slowapi==0.1.9
python-multipart==0.0.9
tenacity==8.2.0
